ocr-services/surya-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
surya-ocr>=0.4.0
Pillow==10.4.0
numpy==1.24.3
python-multipart==0.0.6
torch>=2.0.0
torchvision>=0.15.0
requests==2.32.3
